requirements.txt
fastapi
uvicorn[standard]
pyautogui
mss
Pillow
pytesseract
